docs/irc-whatwg/2016-02-29.html

<!doctype html>
<head>
  <title>#whatwg on 2016-02-29</title>
  <link rel="stylesheet" href="../style.css">
  <script>window.room = "#whatwg";</script><script src="../render-roomlist.js"></script>
  <script src="../logs.js"></script>
</head>
<body><div class="wrapper">
<div class="sidebar">
<div class="title">#whatwg<br>2016-02-29<br><a href="plaintext/">plaintext logs</a></div>
<div class="nav">
<a href="2016-02-28" class="nav-link"><span>prev</span></a>
<a href="2016-03-01" class="nav-link"><span style="float:right">next</span></a>
</div>
    <div class="all-rooms"><noscript>JavaScript is required to load the channel index, but you can go to <a href="../..">the static index</a> directly.</noscript></div>
<div class="footer"><a href="https://github.com/bakkot/matrix-archive-bot">source on github</a></div>
</div>
<div class="rhs">
<div class="rhs-header">
<span id="error" style="color: red; display:none">error</span>
<input type="text" id="query" size=25 placeholder="Search #whatwg">
<a id="search-submit" class="button icon-link" title="Search">
  <svg class="icon" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><path d="M505 442.7L405.3 343c-4.5-4.5-10.6-7-17-7H372c27.6-35.3 44-79.7 44-128C416 93.1 322.9 0 208 0S0 93.1 0 208s93.1 208 208 208c48.3 0 92.7-16.4 128-44v16.3c0 6.4 2.5 12.5 7 17l99.7 99.7c9.4 9.4 24.6 9.4 33.9 0l28.3-28.3c9.4-9.4 9.4-24.6.1-34zM208 336c-70.7 0-128-57.2-128-128 0-70.7 57.2-128 128-128 70.7 0 128 57.2 128 128 0 70.7-57.2 128-128 128z"></path></svg>
</a>
</div>

<div class="log"><table id="log-table"><tbody id="log-tbody">

  <tr class="msg" id="L0"><td class="ts-cell"><a class="ts" href="#L0" alt="Mon Feb 29 2016 03:02:47 GMT-0800 (Pacific Standard Time)">11:02</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">nox: for any future PRs to DOM you should now be able to create branches directly on the main repository (please don&#039;t push to master)</td></tr>
  <tr class="msg" id="L1"><td class="ts-cell"><a class="ts" href="#L1" alt="Mon Feb 29 2016 03:03:26 GMT-0800 (Pacific Standard Time)">11:03</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">nox: the advantage of that is that the person committing (mostly me I guess) can rebase before landing, resulting in automatic closure of the PR and some automatic cross-referencing of the commit to the PR</td></tr>
  <tr class="msg" id="L2"><td class="ts-cell"><a class="ts" href="#L2" alt="Mon Feb 29 2016 03:05:25 GMT-0800 (Pacific Standard Time)">11:05</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">annevk: Ok cool.</td></tr>
  <tr class="msg" id="L3"><td class="ts-cell"><a class="ts" href="#L3" alt="Mon Feb 29 2016 03:07:04 GMT-0800 (Pacific Standard Time)">11:07</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=ritsyy@irc>ritsyy</span>&gt;</div></td><td class="msg-cell">annevk: <a href="https://github.com/whatwg/html/pull/748">https://github.com/whatwg/html/pull/748</a> reviews for this one?, i am unclear about some of the sections</td></tr>
  <tr class="msg" id="L4"><td class="ts-cell"><a class="ts" href="#L4" alt="Mon Feb 29 2016 03:10:16 GMT-0800 (Pacific Standard Time)">11:10</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">ritsyy: looking</td></tr>
  <tr class="msg" id="L5"><td class="ts-cell"><a class="ts" href="#L5" alt="Mon Feb 29 2016 03:14:50 GMT-0800 (Pacific Standard Time)">11:14</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">ritsyy: reviewed</td></tr>
  <tr class="msg" id="L6"><td class="ts-cell"><a class="ts" href="#L6" alt="Mon Feb 29 2016 03:15:39 GMT-0800 (Pacific Standard Time)">11:15</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=ritsyy@irc>ritsyy</span>&gt;</div></td><td class="msg-cell">ritsyy: yes, renaming it, thanks!</td></tr>
  <tr class="msg" id="L7"><td class="ts-cell"><a class="ts" href="#L7" alt="Mon Feb 29 2016 03:15:48 GMT-0800 (Pacific Standard Time)">11:15</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">ritsyy: thanks for the ping, if you pushed that commit a few days ago, remember to say something in the issue, otherwise reviewers won&#039;t get a notification</td></tr>
  <tr class="msg" id="L8"><td class="ts-cell"><a class="ts" href="#L8" alt="Mon Feb 29 2016 03:16:42 GMT-0800 (Pacific Standard Time)">11:16</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=ritsyy@irc>ritsyy</span>&gt;</div></td><td class="msg-cell">annevk: no no i just did the commit, i thought to ping you here, will comment there too after this commit</td></tr>
  <tr class="msg" id="L9"><td class="ts-cell"><a class="ts" href="#L9" alt="Mon Feb 29 2016 03:18:05 GMT-0800 (Pacific Standard Time)">11:18</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">ah okay</td></tr>
  <tr class="msg" id="L10"><td class="ts-cell"><a class="ts" href="#L10" alt="Mon Feb 29 2016 03:19:07 GMT-0800 (Pacific Standard Time)">11:19</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=ralt@irc>ralt</span>&gt;</div></td><td class="msg-cell">annevk: might be worth to switch to gitlab, it has many features that you seem to want (rebase in PRs, protected branches and permissions let some people push to non-protected branches only, etc)</td></tr>
  <tr class="msg" id="L11"><td class="ts-cell"><a class="ts" href="#L11" alt="Mon Feb 29 2016 03:19:14 GMT-0800 (Pacific Standard Time)">11:19</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-3" title=ralt@irc>ralt</span>&gt;</div></td><td class="msg-cell">or at least, look at gitlab :)</td></tr>
  <tr class="msg" id="L12"><td class="ts-cell"><a class="ts" href="#L12" alt="Mon Feb 29 2016 03:22:40 GMT-0800 (Pacific Standard Time)">11:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">ralt: wow cool, I want that, but with the community from GitHub</td></tr>
  <tr class="msg" id="L13"><td class="ts-cell"><a class="ts" href="#L13" alt="Mon Feb 29 2016 03:23:04 GMT-0800 (Pacific Standard Time)">11:23</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">It was already hard to imagine we&#039;re the only ones with these issues</td></tr>
  <tr class="msg" id="L14"><td class="ts-cell"><a class="ts" href="#L14" alt="Mon Feb 29 2016 04:00:53 GMT-0800 (Pacific Standard Time)">12:00</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">annevk: <a href="https://github.com/tabatkins/bikeshed/commit/152f137ad3946c665c8a78648ac0d803369642b5">https://github.com/tabatkins/bikeshed/commit/152f137ad3946c665c8a78648ac0d803369642b5</a></td></tr>
  <tr class="msg" id="L15"><td class="ts-cell"><a class="ts" href="#L15" alt="Mon Feb 29 2016 04:00:57 GMT-0800 (Pacific Standard Time)">12:00</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">That&#039;s the bug I think.</td></tr>
  <tr class="msg" id="L16"><td class="ts-cell"><a class="ts" href="#L16" alt="Mon Feb 29 2016 04:05:57 GMT-0800 (Pacific Standard Time)">12:05</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">annevk: <a href="https://github.com/tabatkins/bikeshed/pull/611">https://github.com/tabatkins/bikeshed/pull/611</a></td></tr>
  <tr class="msg" id="L17"><td class="ts-cell"><a class="ts" href="#L17" alt="Mon Feb 29 2016 05:27:46 GMT-0800 (Pacific Standard Time)">13:27</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">annevk: OOOOOH, I think I got a workaround to unblock DOM.</td></tr>
  <tr class="msg" id="L18"><td class="ts-cell"><a class="ts" href="#L18" alt="Mon Feb 29 2016 05:28:16 GMT-0800 (Pacific Standard Time)">13:28</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Call TabAtkins?</td></tr>
  <tr class="msg" id="L19"><td class="ts-cell"><a class="ts" href="#L19" alt="Mon Feb 29 2016 05:29:06 GMT-0800 (Pacific Standard Time)">13:29</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">annevk: Nah.</td></tr>
  <tr class="msg" id="L20"><td class="ts-cell"><a class="ts" href="#L20" alt="Mon Feb 29 2016 05:29:12 GMT-0800 (Pacific Standard Time)">13:29</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">And I failed. :(</td></tr>
  <tr class="msg" id="L21"><td class="ts-cell"><a class="ts" href="#L21" alt="Mon Feb 29 2016 05:29:28 GMT-0800 (Pacific Standard Time)">13:29</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">annevk: I hoped &amp;equals; would do the trick given it&#039;s just regexps. :P</td></tr>
  <tr class="msg" id="L22"><td class="ts-cell"><a class="ts" href="#L22" alt="Mon Feb 29 2016 05:31:24 GMT-0800 (Pacific Standard Time)">13:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">annevk: We can also do &#039;src= &quot;insanity-wolf&quot;&#039; for now.</td></tr>
  <tr class="msg" id="L23"><td class="ts-cell"><a class="ts" href="#L23" alt="Mon Feb 29 2016 05:31:49 GMT-0800 (Pacific Standard Time)">13:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Nah, let&#039;s wait</td></tr>
  <tr class="msg" id="L24"><td class="ts-cell"><a class="ts" href="#L24" alt="Mon Feb 29 2016 05:32:08 GMT-0800 (Pacific Standard Time)">13:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">TabAtkins discouraged me to work around flaws</td></tr>
  <tr class="msg" id="L25"><td class="ts-cell"><a class="ts" href="#L25" alt="Mon Feb 29 2016 05:33:00 GMT-0800 (Pacific Standard Time)">13:33</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">K.</td></tr>
  <tr class="msg" id="L26"><td class="ts-cell"><a class="ts" href="#L26" alt="Mon Feb 29 2016 07:20:09 GMT-0800 (Pacific Standard Time)">15:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">We should set up CI for the bikeshed specs so that (a) we don&#039;t have to check in generated code; (b) we can pin to a specific Bikeshed version.</td></tr>
  <tr class="msg" id="L27"><td class="ts-cell"><a class="ts" href="#L27" alt="Mon Feb 29 2016 07:20:33 GMT-0800 (Pacific Standard Time)">15:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">I envision a metadata file in the repo (maybe .travis.yml, maybe something else) that contains the Bikeshed commit hash that the spec is meant to be built with</td></tr>
  <tr class="msg" id="L28"><td class="ts-cell"><a class="ts" href="#L28" alt="Mon Feb 29 2016 07:20:37 GMT-0800 (Pacific Standard Time)">15:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">And you can update that every once in a while</td></tr>
  <tr class="msg" id="L29"><td class="ts-cell"><a class="ts" href="#L29" alt="Mon Feb 29 2016 07:32:51 GMT-0800 (Pacific Standard Time)">15:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">I would not mind such a setup</td></tr>
  <tr class="msg" id="L30"><td class="ts-cell"><a class="ts" href="#L30" alt="Mon Feb 29 2016 07:39:03 GMT-0800 (Pacific Standard Time)">15:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-2" title=gsnedders@irc>gsnedders</span>&gt;</div></td><td class="msg-cell">I guess the important question is whether we have any desire to review BS output. Probably not really, because it&#039;s just too noisy.</td></tr>
  <tr class="msg" id="L31"><td class="ts-cell"><a class="ts" href="#L31" alt="Mon Feb 29 2016 07:45:32 GMT-0800 (Pacific Standard Time)">15:45</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">gsnedders: it&#039;s somewhat useful to do, especially since I often type &lt;span&gt; instead of &lt;a&gt;, but it&#039;s very noisy indeed</td></tr>
  <tr class="msg" id="L32"><td class="ts-cell"><a class="ts" href="#L32" alt="Mon Feb 29 2016 08:02:58 GMT-0800 (Pacific Standard Time)">16:02</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">TIL about the word &quot;excrescence&quot;</td></tr>
  <tr class="msg" id="L33"><td class="ts-cell"><a class="ts" href="#L33" alt="Mon Feb 29 2016 08:16:36 GMT-0800 (Pacific Standard Time)">16:16</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">1 of 1 matches</td></tr>
  <tr class="msg" id="L34"><td class="ts-cell"><a class="ts" href="#L34" alt="Mon Feb 29 2016 08:23:33 GMT-0800 (Pacific Standard Time)">16:23</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">foolip++</td></tr>
  <tr class="msg" id="L35"><td class="ts-cell"><a class="ts" href="#L35" alt="Mon Feb 29 2016 08:24:51 GMT-0800 (Pacific Standard Time)">16:24</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-11" title=philipj@irc>philipj</span>&gt;</div></td><td class="msg-cell">Ms2ger: I tried quite hard to find something to complain about, but the truth is you write tests almost exactly like I would aspire to :)</td></tr>
  <tr class="msg" id="L36"><td class="ts-cell"><a class="ts" href="#L36" alt="Mon Feb 29 2016 08:26:02 GMT-0800 (Pacific Standard Time)">16:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-11" title=philipj@irc>philipj</span>&gt;</div></td><td class="msg-cell">Ms2ger: if you&#039;re in a good mood, the top two in <a href="https://github.com/w3c/web-platform-tests/pulls/foolip">https://github.com/w3c/web-platform-tests/pulls/foolip</a> need review</td></tr>
  <tr class="msg" id="L37"><td class="ts-cell"><a class="ts" href="#L37" alt="Mon Feb 29 2016 08:26:27 GMT-0800 (Pacific Standard Time)">16:26</a></td><td class="nick-cell"><div class="m-ov"><span class="nick nick-11" title=philipj@irc>philipj</span></div></td><td class="msg-cell">does actually know what mood has to do with it...</td></tr>
  <tr class="msg" id="L38"><td class="ts-cell"><a class="ts" href="#L38" alt="Mon Feb 29 2016 08:26:32 GMT-0800 (Pacific Standard Time)">16:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-11" title=philipj@irc>philipj</span>&gt;</div></td><td class="msg-cell">doesn&#039;t!</td></tr>
  <tr class="msg" id="L39"><td class="ts-cell"><a class="ts" href="#L39" alt="Mon Feb 29 2016 08:29:51 GMT-0800 (Pacific Standard Time)">16:29</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">philipj, r+ and r+ with rebase and nit</td></tr>
  <tr class="msg" id="L40"><td class="ts-cell"><a class="ts" href="#L40" alt="Mon Feb 29 2016 08:35:15 GMT-0800 (Pacific Standard Time)">16:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">While we&#039;re talking tests... is <a href="https://github.com/yaycmyk/jsdom/blob/label-input-activation/test/web-platform-tests/to-upstream/html/semantics/forms/the-label-element/proxy-click-to-associated-element.html#L24-L29">https://github.com/yaycmyk/jsdom/blob/label-input-activation/test/web-platform-tests/to-upstream/html/semantics/forms/the-label-element/proxy-click-to-associated-element.html#L24-L29</a> the correct pattern for testing events, or should we be using EventWatcher maybe, or...?</td></tr>
  <tr class="msg" id="L41"><td class="ts-cell"><a class="ts" href="#L41" alt="Mon Feb 29 2016 08:35:39 GMT-0800 (Pacific Standard Time)">16:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">EventWatcher seems pretty heavyweight for single events I guess</td></tr>
  <tr class="msg" id="L42"><td class="ts-cell"><a class="ts" href="#L42" alt="Mon Feb 29 2016 08:36:05 GMT-0800 (Pacific Standard Time)">16:36</a></td><td class="nick-cell"><div class="m-ov"><span class="nick nick-12" title=annevk@irc>annevk</span></div></td><td class="msg-cell">wonders if nox has studied Shadow DOM yet</td></tr>
  <tr class="msg" id="L43"><td class="ts-cell"><a class="ts" href="#L43" alt="Mon Feb 29 2016 08:36:13 GMT-0800 (Pacific Standard Time)">16:36</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">Not yet.</td></tr>
  <tr class="msg" id="L44"><td class="ts-cell"><a class="ts" href="#L44" alt="Mon Feb 29 2016 08:36:24 GMT-0800 (Pacific Standard Time)">16:36</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">But given I implemented &lt;template&gt;, I probably will at some point.</td></tr>
  <tr class="msg" id="L45"><td class="ts-cell"><a class="ts" href="#L45" alt="Mon Feb 29 2016 08:36:50 GMT-0800 (Pacific Standard Time)">16:36</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Domenic, assert_true(true); is dumb</td></tr>
  <tr class="msg" id="L46"><td class="ts-cell"><a class="ts" href="#L46" alt="Mon Feb 29 2016 08:37:05 GMT-0800 (Pacific Standard Time)">16:37</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">Ms2ger: just do t.done()?</td></tr>
  <tr class="msg" id="L47"><td class="ts-cell"><a class="ts" href="#L47" alt="Mon Feb 29 2016 08:37:15 GMT-0800 (Pacific Standard Time)">16:37</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Domenic, input.addEventListener(&quot;click&quot;, t.step_func_done()) is probably what I&#039;d use</td></tr>
  <tr class="msg" id="L48"><td class="ts-cell"><a class="ts" href="#L48" alt="Mon Feb 29 2016 08:37:53 GMT-0800 (Pacific Standard Time)">16:37</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">Ms2ger: hmm, that&#039;s not documented in <a href="http://testthewebforward.org/docs/testharness-library.html#promise-tests">http://testthewebforward.org/docs/testharness-library.html#promise-tests</a></td></tr>
  <tr class="msg" id="L49"><td class="ts-cell"><a class="ts" href="#L49" alt="Mon Feb 29 2016 08:38:27 GMT-0800 (Pacific Standard Time)">16:38</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">You&#039;re correct</td></tr>
  <tr class="msg" id="L50"><td class="ts-cell"><a class="ts" href="#L50" alt="Mon Feb 29 2016 08:39:07 GMT-0800 (Pacific Standard Time)">16:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">I&#039;ll see if I can pull off a PR from the web interface</td></tr>
  <tr class="msg" id="L51"><td class="ts-cell"><a class="ts" href="#L51" alt="Mon Feb 29 2016 08:39:28 GMT-0800 (Pacific Standard Time)">16:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: ETA on structured clone?</td></tr>
  <tr class="msg" id="L52"><td class="ts-cell"><a class="ts" href="#L52" alt="Mon Feb 29 2016 08:39:36 GMT-0800 (Pacific Standard Time)">16:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-11" title=philipj@irc>philipj</span>&gt;</div></td><td class="msg-cell">Ms2ger: thanks!</td></tr>
  <tr class="msg" id="L53"><td class="ts-cell"><a class="ts" href="#L53" alt="Mon Feb 29 2016 08:39:37 GMT-0800 (Pacific Standard Time)">16:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: within the hour</td></tr>
  <tr class="msg" id="L54"><td class="ts-cell"><a class="ts" href="#L54" alt="Mon Feb 29 2016 08:39:48 GMT-0800 (Pacific Standard Time)">16:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">Ms2ger: oh lol I see why they did assert_true. The docs told them to.</td></tr>
  <tr class="msg" id="L55"><td class="ts-cell"><a class="ts" href="#L55" alt="Mon Feb 29 2016 08:39:58 GMT-0800 (Pacific Standard Time)">16:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">Ms2ger: &quot;The above example can be rewritten as:&quot;</td></tr>
  <tr class="msg" id="L56"><td class="ts-cell"><a class="ts" href="#L56" alt="Mon Feb 29 2016 08:40:25 GMT-0800 (Pacific Standard Time)">16:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: I want to make the change I suggested in the PR, but other than that I think it&#039;s good, but who knows, you might find something again 😃</td></tr>
  <tr class="msg" id="L57"><td class="ts-cell"><a class="ts" href="#L57" alt="Mon Feb 29 2016 08:40:25 GMT-0800 (Pacific Standard Time)">16:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Huh</td></tr>
  <tr class="msg" id="L58"><td class="ts-cell"><a class="ts" href="#L58" alt="Mon Feb 29 2016 08:40:36 GMT-0800 (Pacific Standard Time)">16:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: in the PR thread*</td></tr>
  <tr class="msg" id="L59"><td class="ts-cell"><a class="ts" href="#L59" alt="Mon Feb 29 2016 08:41:28 GMT-0800 (Pacific Standard Time)">16:41</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">I will change it to assert_equals(1+1, 2)</td></tr>
  <tr class="msg" id="L60"><td class="ts-cell"><a class="ts" href="#L60" alt="Mon Feb 29 2016 08:48:08 GMT-0800 (Pacific Standard Time)">16:48</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Domenic, are you updating the docs?</td></tr>
  <tr class="msg" id="L61"><td class="ts-cell"><a class="ts" href="#L61" alt="Mon Feb 29 2016 08:48:45 GMT-0800 (Pacific Standard Time)">16:48</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">Ms2ger: <a href="https://github.com/w3c/testharness.js/pull/186">https://github.com/w3c/testharness.js/pull/186</a></td></tr>
  <tr class="msg" id="L62"><td class="ts-cell"><a class="ts" href="#L62" alt="Mon Feb 29 2016 08:51:11 GMT-0800 (Pacific Standard Time)">16:51</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Domenic, reviewed</td></tr>
  <tr class="msg" id="L63"><td class="ts-cell"><a class="ts" href="#L63" alt="Mon Feb 29 2016 08:51:30 GMT-0800 (Pacific Standard Time)">16:51</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">Ms2ger: dang, now I have to go make changes using not the web interface :P. Will fix after reviewing this HTML PR...</td></tr>
  <tr class="msg" id="L64"><td class="ts-cell"><a class="ts" href="#L64" alt="Mon Feb 29 2016 08:52:11 GMT-0800 (Pacific Standard Time)">16:52</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-4" title=Ms2ger@irc>Ms2ger</span>&gt;</div></td><td class="msg-cell">Domenic, I&#039;ll review again tomorrow if nobody beats me to it :)</td></tr>
  <tr class="msg" id="L65"><td class="ts-cell"><a class="ts" href="#L65" alt="Mon Feb 29 2016 09:10:26 GMT-0800 (Pacific Standard Time)">17:10</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">annevk: I&#039;m working on it this morning, don&#039;t worry.</td></tr>
  <tr class="msg" id="L66"><td class="ts-cell"><a class="ts" href="#L66" alt="Mon Feb 29 2016 09:12:52 GMT-0800 (Pacific Standard Time)">17:12</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">nox&#039;s PR seems hacky but workable, I&#039;ll merge in a few minutes.</td></tr>
  <tr class="msg" id="L67"><td class="ts-cell"><a class="ts" href="#L67" alt="Mon Feb 29 2016 09:13:01 GMT-0800 (Pacific Standard Time)">17:13</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">TabAtkins: It doesn&#039;t pass tests.</td></tr>
  <tr class="msg" id="L68"><td class="ts-cell"><a class="ts" href="#L68" alt="Mon Feb 29 2016 09:13:16 GMT-0800 (Pacific Standard Time)">17:13</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">TabAtkins: I&#039;m not sure what&#039;s that thing is supposed to match btw, because I don&#039;t see =type= in that shorthands file.</td></tr>
  <tr class="msg" id="L69"><td class="ts-cell"><a class="ts" href="#L69" alt="Mon Feb 29 2016 09:13:22 GMT-0800 (Pacific Standard Time)">17:13</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">Yeah, that&#039;s why I can&#039;t merge from my phone.</td></tr>
  <tr class="msg" id="L70"><td class="ts-cell"><a class="ts" href="#L70" alt="Mon Feb 29 2016 09:13:46 GMT-0800 (Pacific Standard Time)">17:13</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">Haven&#039;t updated the docs yet, was waiting for any problems to shake out. 😀</td></tr>
  <tr class="msg" id="L71"><td class="ts-cell"><a class="ts" href="#L71" alt="Mon Feb 29 2016 09:22:02 GMT-0800 (Pacific Standard Time)">17:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">ugh</td></tr>
  <tr class="msg" id="L72"><td class="ts-cell"><a class="ts" href="#L72" alt="Mon Feb 29 2016 09:22:16 GMT-0800 (Pacific Standard Time)">17:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">TabAtkins: it basically delayed fixing stuff in DOM by two days, that&#039;s not a great way to operate</td></tr>
  <tr class="msg" id="L73"><td class="ts-cell"><a class="ts" href="#L73" alt="Mon Feb 29 2016 09:22:23 GMT-0800 (Pacific Standard Time)">17:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">TabAtkins: please fix the way you develop bikeshed</td></tr>
  <tr class="msg" id="L74"><td class="ts-cell"><a class="ts" href="#L74" alt="Mon Feb 29 2016 09:26:26 GMT-0800 (Pacific Standard Time)">17:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: const a = []; Object.defineProperty(a, &quot;length&quot;, { get() { return 5; } }) throws</td></tr>
  <tr class="msg" id="L75"><td class="ts-cell"><a class="ts" href="#L75" alt="Mon Feb 29 2016 09:26:48 GMT-0800 (Pacific Standard Time)">17:26</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: I based my code on similar-looking code in ECMAScript</td></tr>
  <tr class="msg" id="L76"><td class="ts-cell"><a class="ts" href="#L76" alt="Mon Feb 29 2016 09:27:06 GMT-0800 (Pacific Standard Time)">17:27</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: hmm, I don&#039;t think it should, per spec...</td></tr>
  <tr class="msg" id="L77"><td class="ts-cell"><a class="ts" href="#L77" alt="Mon Feb 29 2016 09:27:35 GMT-0800 (Pacific Standard Time)">17:27</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: ah I see, length is non-configurable...</td></tr>
  <tr class="msg" id="L78"><td class="ts-cell"><a class="ts" href="#L78" alt="Mon Feb 29 2016 09:28:03 GMT-0800 (Pacific Standard Time)">17:28</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: OK, that part looks good</td></tr>
  <tr class="msg" id="L79"><td class="ts-cell"><a class="ts" href="#L79" alt="Mon Feb 29 2016 09:32:01 GMT-0800 (Pacific Standard Time)">17:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: should we go through the placeholder stuff here?</td></tr>
  <tr class="msg" id="L80"><td class="ts-cell"><a class="ts" href="#L80" alt="Mon Feb 29 2016 09:32:06 GMT-0800 (Pacific Standard Time)">17:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: or is it clear now?</td></tr>
  <tr class="msg" id="L81"><td class="ts-cell"><a class="ts" href="#L81" alt="Mon Feb 29 2016 09:32:24 GMT-0800 (Pacific Standard Time)">17:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: it is definitely not clear, but I will try re-reading it with your new explanation in mind and suggesting better text...</td></tr>
  <tr class="msg" id="L82"><td class="ts-cell"><a class="ts" href="#L82" alt="Mon Feb 29 2016 09:32:48 GMT-0800 (Pacific Standard Time)">17:32</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: this is basically what the old text had too</td></tr>
  <tr class="msg" id="L83"><td class="ts-cell"><a class="ts" href="#L83" alt="Mon Feb 29 2016 09:33:01 GMT-0800 (Pacific Standard Time)">17:33</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: though replacing the placeholder with the transferred object lived in postMessage()</td></tr>
  <tr class="msg" id="L84"><td class="ts-cell"><a class="ts" href="#L84" alt="Mon Feb 29 2016 09:33:05 GMT-0800 (Pacific Standard Time)">17:33</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: well, I didn&#039;t review the old text :)</td></tr>
  <tr class="msg" id="L85"><td class="ts-cell"><a class="ts" href="#L85" alt="Mon Feb 29 2016 09:33:37 GMT-0800 (Pacific Standard Time)">17:33</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: sure, but basically a lot of improvements are being blocked on this now</td></tr>
  <tr class="msg" id="L86"><td class="ts-cell"><a class="ts" href="#L86" alt="Mon Feb 29 2016 09:34:34 GMT-0800 (Pacific Standard Time)">17:34</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: could it be rephrased as &quot;Everywhere placeholderResult shows up inside clone, replace it with transferResult&quot;?</td></tr>
  <tr class="msg" id="L87"><td class="ts-cell"><a class="ts" href="#L87" alt="Mon Feb 29 2016 09:34:35 GMT-0800 (Pacific Standard Time)">17:34</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: basically if you have an object-graph in your engine, you want to take the placeholder object and replace it directly with the transferred object</td></tr>
  <tr class="msg" id="L88"><td class="ts-cell"><a class="ts" href="#L88" alt="Mon Feb 29 2016 09:34:58 GMT-0800 (Pacific Standard Time)">17:34</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: right but e.g. does that mean crawling the graph with GetOwnProperties etc.?</td></tr>
  <tr class="msg" id="L89"><td class="ts-cell"><a class="ts" href="#L89" alt="Mon Feb 29 2016 09:35:13 GMT-0800 (Pacific Standard Time)">17:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: no, because then if it didn&#039;t show up inside clone it wouldn&#039;t be replaced</td></tr>
  <tr class="msg" id="L90"><td class="ts-cell"><a class="ts" href="#L90" alt="Mon Feb 29 2016 09:35:48 GMT-0800 (Pacific Standard Time)">17:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: that is the main part of my question. where else can it show up.</td></tr>
  <tr class="msg" id="L91"><td class="ts-cell"><a class="ts" href="#L91" alt="Mon Feb 29 2016 09:35:52 GMT-0800 (Pacific Standard Time)">17:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: I&#039;m not sure what GetOwnProperties is or how it relates</td></tr>
  <tr class="msg" id="L92"><td class="ts-cell"><a class="ts" href="#L92" alt="Mon Feb 29 2016 09:36:06 GMT-0800 (Pacific Standard Time)">17:36</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: it&#039;s returned in [[TransferList]]</td></tr>
  <tr class="msg" id="L93"><td class="ts-cell"><a class="ts" href="#L93" alt="Mon Feb 29 2016 09:36:35 GMT-0800 (Pacific Standard Time)">17:36</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: hmm, maybe it cannot show up there, those don&#039;t use memory at the moment</td></tr>
  <tr class="msg" id="L94"><td class="ts-cell"><a class="ts" href="#L94" alt="Mon Feb 29 2016 09:36:48 GMT-0800 (Pacific Standard Time)">17:36</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: right that&#039;s step 5.4. But in 5.3, which we are discussing, it isn&#039;t in outputTransferList</td></tr>
  <tr class="msg" id="L95"><td class="ts-cell"><a class="ts" href="#L95" alt="Mon Feb 29 2016 09:37:36 GMT-0800 (Pacific Standard Time)">17:37</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: I was thinking that objects ending up in [[TransferList]] could have references, but that&#039;s currently not possible</td></tr>
  <tr class="msg" id="L96"><td class="ts-cell"><a class="ts" href="#L96" alt="Mon Feb 29 2016 09:37:54 GMT-0800 (Pacific Standard Time)">17:37</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: that might change though if we start passing memory to [[Transfer]] in the future for some new use case</td></tr>
  <tr class="msg" id="L97"><td class="ts-cell"><a class="ts" href="#L97" alt="Mon Feb 29 2016 09:38:04 GMT-0800 (Pacific Standard Time)">17:38</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: let&#039;s worry about that then?</td></tr>
  <tr class="msg" id="L98"><td class="ts-cell"><a class="ts" href="#L98" alt="Mon Feb 29 2016 09:38:56 GMT-0800 (Pacific Standard Time)">17:38</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: sure, I can add a comment under [[Transfer]] similar to what I have under [[Clone]]</td></tr>
  <tr class="msg" id="L99"><td class="ts-cell"><a class="ts" href="#L99" alt="Mon Feb 29 2016 09:39:12 GMT-0800 (Pacific Standard Time)">17:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">What kind of comment?</td></tr>
  <tr class="msg" id="L100"><td class="ts-cell"><a class="ts" href="#L100" alt="Mon Feb 29 2016 09:39:16 GMT-0800 (Pacific Standard Time)">17:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">Oh you mean &lt;!-- comment</td></tr>
  <tr class="msg" id="L101"><td class="ts-cell"><a class="ts" href="#L101" alt="Mon Feb 29 2016 09:39:28 GMT-0800 (Pacific Standard Time)">17:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">&lt;!-- if we add memory, we need to change so and so --&gt;</td></tr>
  <tr class="msg" id="L102"><td class="ts-cell"><a class="ts" href="#L102" alt="Mon Feb 29 2016 09:39:46 GMT-0800 (Pacific Standard Time)">17:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">Thanks.</td></tr>
  <tr class="msg" id="L103"><td class="ts-cell"><a class="ts" href="#L103" alt="Mon Feb 29 2016 09:39:57 GMT-0800 (Pacific Standard Time)">17:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">This pushes the major imprecision into the single phrase &quot;inside clone&quot;, which we can live with</td></tr>
  <tr class="msg" id="L104"><td class="ts-cell"><a class="ts" href="#L104" alt="Mon Feb 29 2016 09:40:19 GMT-0800 (Pacific Standard Time)">17:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">Ideally we&#039;d specify exactly how you go crawling the object graph and replacing these placeholders. Or figure out what implementations actually do.</td></tr>
  <tr class="msg" id="L105"><td class="ts-cell"><a class="ts" href="#L105" alt="Mon Feb 29 2016 09:40:21 GMT-0800 (Pacific Standard Time)">17:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">All right, nox&#039;s fix worked, dom builds cleanly now.</td></tr>
  <tr class="msg" id="L106"><td class="ts-cell"><a class="ts" href="#L106" alt="Mon Feb 29 2016 09:40:45 GMT-0800 (Pacific Standard Time)">17:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">I see, I thought that was more or less a given since we created placeholder objects inside the StructuredCloneWithTransfer algorithm and that algorithm only ever returns clone/transferList or an exception</td></tr>
  <tr class="msg" id="L107"><td class="ts-cell"><a class="ts" href="#L107" alt="Mon Feb 29 2016 09:42:09 GMT-0800 (Pacific Standard Time)">17:42</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: ^ still not entirely clear to me why you think it&#039;s unclear now, but I guess I can add &quot;in /clone/&quot;</td></tr>
  <tr class="msg" id="L108"><td class="ts-cell"><a class="ts" href="#L108" alt="Mon Feb 29 2016 09:43:21 GMT-0800 (Pacific Standard Time)">17:43</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: I am trying to figure out if there are observable consequences to different ways of doing this replacement (because there are many: different ways of graph traversal; different ways of getting the list of properties; etc.) but I am not sure yet one way or the other.</td></tr>
  <tr class="msg" id="L109"><td class="ts-cell"><a class="ts" href="#L109" alt="Mon Feb 29 2016 09:43:57 GMT-0800 (Pacific Standard Time)">17:43</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">I suspect this actually happens at a much lower level than getting properties in engines</td></tr>
  <tr class="msg" id="L110"><td class="ts-cell"><a class="ts" href="#L110" alt="Mon Feb 29 2016 09:44:35 GMT-0800 (Pacific Standard Time)">17:44</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">They have these placeholder objects allocated with some pointer, and they just make that point someplace else</td></tr>
  <tr class="msg" id="L111"><td class="ts-cell"><a class="ts" href="#L111" alt="Mon Feb 29 2016 09:44:47 GMT-0800 (Pacific Standard Time)">17:44</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">Right, I guess so.</td></tr>
  <tr class="msg" id="L112"><td class="ts-cell"><a class="ts" href="#L112" alt="Mon Feb 29 2016 09:44:51 GMT-0800 (Pacific Standard Time)">17:44</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">nox: Unfortunately using entities doesn&#039;t avoid the shorthands, as all but a few run over the parsed DOM, not raw text, so they see the entity already substituted. Haven&#039;t come up with a good way around that yet. :/</td></tr>
  <tr class="msg" id="L113"><td class="ts-cell"><a class="ts" href="#L113" alt="Mon Feb 29 2016 09:45:32 GMT-0800 (Pacific Standard Time)">17:45</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-15" title=nox@irc>nox</span>&gt;</div></td><td class="msg-cell">TabAtkins: Yeah understood that afterwards. :) I saw regexps and just thought bikeshed was crazier than I had thought initially. Sorry for that. :P</td></tr>
  <tr class="msg" id="L114"><td class="ts-cell"><a class="ts" href="#L114" alt="Mon Feb 29 2016 09:46:03 GMT-0800 (Pacific Standard Time)">17:46</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: it would only be observable if you could get cause code to run in targetRealm while cloning, which should be impossible</td></tr>
  <tr class="msg" id="L115"><td class="ts-cell"><a class="ts" href="#L115" alt="Mon Feb 29 2016 09:46:35 GMT-0800 (Pacific Standard Time)">17:46</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: I agree that it would be nice if we had better language to define this, but I don&#039;t think ECMAScript gives low-level access to allocated objects</td></tr>
  <tr class="msg" id="L116"><td class="ts-cell"><a class="ts" href="#L116" alt="Mon Feb 29 2016 09:47:11 GMT-0800 (Pacific Standard Time)">17:47</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">nox: I&#039;d actually thought of the case that DOM hit, but remembered that it wouldn&#039;t be hit by normal attributes. I did not consider code examples. ^_^</td></tr>
  <tr class="msg" id="L117"><td class="ts-cell"><a class="ts" href="#L117" alt="Mon Feb 29 2016 09:47:34 GMT-0800 (Pacific Standard Time)">17:47</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: maybe I should ask in #jslang if anyone has ideas</td></tr>
  <tr class="msg" id="L118"><td class="ts-cell"><a class="ts" href="#L118" alt="Mon Feb 29 2016 09:47:46 GMT-0800 (Pacific Standard Time)">17:47</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">Anyway, DOM is being put into the integration suite now, and I&#039;m gonna put a bunch more in too, so it&#039;s unlikely I&#039;ll break anything without noticing again.</td></tr>
  <tr class="msg" id="L119"><td class="ts-cell"><a class="ts" href="#L119" alt="Mon Feb 29 2016 09:51:32 GMT-0800 (Pacific Standard Time)">17:51</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">TabAtkins: thank you, <a href="http://dom.bs">dom.bs</a> and <a href="http://url.bs">url.bs</a> are the main pain points</td></tr>
  <tr class="msg" id="L120"><td class="ts-cell"><a class="ts" href="#L120" alt="Mon Feb 29 2016 09:51:39 GMT-0800 (Pacific Standard Time)">17:51</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">TabAtkins: <a href="http://notifications.bs">notifications.bs</a> seems to get by, mostly</td></tr>
  <tr class="msg" id="L121"><td class="ts-cell"><a class="ts" href="#L121" alt="Mon Feb 29 2016 09:52:09 GMT-0800 (Pacific Standard Time)">17:52</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: asked there, but no high hopes</td></tr>
  <tr class="msg" id="L122"><td class="ts-cell"><a class="ts" href="#L122" alt="Mon Feb 29 2016 09:52:14 GMT-0800 (Pacific Standard Time)">17:52</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">annevk: kk, I&#039;ll put <a href="http://url.bs">url.bs</a> in too. No reason not to, testing is already slow enough that slowing it down by another minute or so is fine.</td></tr>
  <tr class="msg" id="L123"><td class="ts-cell"><a class="ts" href="#L123" alt="Mon Feb 29 2016 09:52:20 GMT-0800 (Pacific Standard Time)">17:52</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: just clarifying that it&#039;s scoped to clone is enough for now for me</td></tr>
  <tr class="msg" id="L124"><td class="ts-cell"><a class="ts" href="#L124" alt="Mon Feb 29 2016 09:52:45 GMT-0800 (Pacific Standard Time)">17:52</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">TabAtkins: you might want to improve those build times :-P</td></tr>
  <tr class="msg" id="L125"><td class="ts-cell"><a class="ts" href="#L125" alt="Mon Feb 29 2016 09:53:09 GMT-0800 (Pacific Standard Time)">17:53</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">Man don&#039;t I know it.  Got a few things lined up to work on for that, then gonna dive in and see what else is slowing things down.</td></tr>
  <tr class="msg" id="L126"><td class="ts-cell"><a class="ts" href="#L126" alt="Mon Feb 29 2016 09:53:30 GMT-0800 (Pacific Standard Time)">17:53</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">Got a decent profiler already hooked up, so it&#039;s just a matter of putting in the time.</td></tr>
  <tr class="msg" id="L127"><td class="ts-cell"><a class="ts" href="#L127" alt="Mon Feb 29 2016 10:13:59 GMT-0800 (Pacific Standard Time)">18:13</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=IZh@irc>IZh</span>&gt;</div></td><td class="msg-cell">Hi. Is this a valid link? &lt;a href=#shared-internal-slot:-[[crossoriginpropertydescriptormap]]&gt;...&lt;/a&gt;</td></tr>
  <tr class="msg" id="L128"><td class="ts-cell"><a class="ts" href="#L128" alt="Mon Feb 29 2016 10:15:25 GMT-0800 (Pacific Standard Time)">18:15</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">think so</td></tr>
  <tr class="msg" id="L129"><td class="ts-cell"><a class="ts" href="#L129" alt="Mon Feb 29 2016 10:15:52 GMT-0800 (Pacific Standard Time)">18:15</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=IZh@irc>IZh</span>&gt;</div></td><td class="msg-cell">annevk: Shouldn&#039;t square brackets be %-encoded?</td></tr>
  <tr class="msg" id="L130"><td class="ts-cell"><a class="ts" href="#L130" alt="Mon Feb 29 2016 10:16:57 GMT-0800 (Pacific Standard Time)">18:16</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">IZh: I guess per <a href="https://url.spec.whatwg.org/#syntax-url-fragment">https://url.spec.whatwg.org/#syntax-url-fragment</a> they should</td></tr>
  <tr class="msg" id="L131"><td class="ts-cell"><a class="ts" href="#L131" alt="Mon Feb 29 2016 10:17:24 GMT-0800 (Pacific Standard Time)">18:17</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">IZh: that might be hard to fix though</td></tr>
  <tr class="msg" id="L132"><td class="ts-cell"><a class="ts" href="#L132" alt="Mon Feb 29 2016 10:17:50 GMT-0800 (Pacific Standard Time)">18:17</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=IZh@irc>IZh</span>&gt;</div></td><td class="msg-cell">annevk: Well, the spec has only one wrong anchor.</td></tr>
  <tr class="msg" id="L133"><td class="ts-cell"><a class="ts" href="#L133" alt="Mon Feb 29 2016 10:18:20 GMT-0800 (Pacific Standard Time)">18:18</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">IZh: it&#039;s generated, that&#039;s the problem</td></tr>
  <tr class="msg" id="L134"><td class="ts-cell"><a class="ts" href="#L134" alt="Mon Feb 29 2016 10:18:35 GMT-0800 (Pacific Standard Time)">18:18</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">IZh: although I suppose we could just override the id of that header and forget about it</td></tr>
  <tr class="msg" id="L135"><td class="ts-cell"><a class="ts" href="#L135" alt="Mon Feb 29 2016 10:19:29 GMT-0800 (Pacific Standard Time)">18:19</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=IZh@irc>IZh</span>&gt;</div></td><td class="msg-cell">annevk: But why not to fix escaping code? I suppose it must already %-encode non-ASCII symbols. Why not just add the square brackets to the set?</td></tr>
  <tr class="msg" id="L136"><td class="ts-cell"><a class="ts" href="#L136" alt="Mon Feb 29 2016 10:19:45 GMT-0800 (Pacific Standard Time)">18:19</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">annevk: I&#039;m not url-encoding anchors in several spots, I&#039;m happy to go through and fix that.</td></tr>
  <tr class="msg" id="L137"><td class="ts-cell"><a class="ts" href="#L137" alt="Mon Feb 29 2016 10:20:56 GMT-0800 (Pacific Standard Time)">18:20</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">(I haven&#039;t needed to be precise with that, because Bikeshed only generates IDs that are url-safe without any escaping needed.)</td></tr>
  <tr class="msg" id="L138"><td class="ts-cell"><a class="ts" href="#L138" alt="Mon Feb 29 2016 10:22:46 GMT-0800 (Pacific Standard Time)">18:22</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">This is Wattsi though :)</td></tr>
  <tr class="msg" id="L139"><td class="ts-cell"><a class="ts" href="#L139" alt="Mon Feb 29 2016 10:24:52 GMT-0800 (Pacific Standard Time)">18:24</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=IZh@irc>IZh</span>&gt;</div></td><td class="msg-cell">It seems that https is not supported on IPv6 address of <a href="http://whatwg.org">whatwg.org</a>.</td></tr>
  <tr class="msg" id="L140"><td class="ts-cell"><a class="ts" href="#L140" alt="Mon Feb 29 2016 10:27:13 GMT-0800 (Pacific Standard Time)">18:27</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">IZh: as I said, that might be tricky</td></tr>
  <tr class="msg" id="L141"><td class="ts-cell"><a class="ts" href="#L141" alt="Mon Feb 29 2016 10:27:24 GMT-0800 (Pacific Standard Time)">18:27</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">IZh: file an issue against whatwg/wattsi</td></tr>
  <tr class="msg" id="L142"><td class="ts-cell"><a class="ts" href="#L142" alt="Mon Feb 29 2016 10:27:30 GMT-0800 (Pacific Standard Time)">18:27</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">IZh: or a PR 😃</td></tr>
  <tr class="msg" id="L143"><td class="ts-cell"><a class="ts" href="#L143" alt="Mon Feb 29 2016 10:30:03 GMT-0800 (Pacific Standard Time)">18:30</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: if you can take another look now maybe we can land it today</td></tr>
  <tr class="msg" id="L144"><td class="ts-cell"><a class="ts" href="#L144" alt="Mon Feb 29 2016 10:31:31 GMT-0800 (Pacific Standard Time)">18:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=IZh@irc>IZh</span>&gt;</div></td><td class="msg-cell">annevk: Filed #22.</td></tr>
  <tr class="msg" id="L145"><td class="ts-cell"><a class="ts" href="#L145" alt="Mon Feb 29 2016 10:33:21 GMT-0800 (Pacific Standard Time)">18:33</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">IZh: thank you, is this breaking anything for you or just something you noticed?</td></tr>
  <tr class="msg" id="L146"><td class="ts-cell"><a class="ts" href="#L146" alt="Mon Feb 29 2016 10:34:20 GMT-0800 (Pacific Standard Time)">18:34</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=IZh@irc>IZh</span>&gt;</div></td><td class="msg-cell">annevk: Just something another validator is barking on. :-)</td></tr>
  <tr class="msg" id="L147"><td class="ts-cell"><a class="ts" href="#L147" alt="Mon Feb 29 2016 10:34:51 GMT-0800 (Pacific Standard Time)">18:34</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=IZh@irc>IZh</span>&gt;</div></td><td class="msg-cell">annevk: I&#039;m periodically validating the spec. :-)</td></tr>
  <tr class="msg" id="L148"><td class="ts-cell"><a class="ts" href="#L148" alt="Mon Feb 29 2016 10:35:32 GMT-0800 (Pacific Standard Time)">18:35</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=IZh@irc>IZh</span>&gt;</div></td><td class="msg-cell">Because the spec has to conform to itself.</td></tr>
  <tr class="msg" id="L149"><td class="ts-cell"><a class="ts" href="#L149" alt="Mon Feb 29 2016 10:37:36 GMT-0800 (Pacific Standard Time)">18:37</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: would &quot;across event loops, i.e., ...&quot; work?</td></tr>
  <tr class="msg" id="L150"><td class="ts-cell"><a class="ts" href="#L150" alt="Mon Feb 29 2016 10:37:42 GMT-0800 (Pacific Standard Time)">18:37</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: I mainly don&#039;t like the &quot;:&quot;</td></tr>
  <tr class="msg" id="L151"><td class="ts-cell"><a class="ts" href="#L151" alt="Mon Feb 29 2016 10:38:38 GMT-0800 (Pacific Standard Time)">18:38</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">Yeah it would. I like the colon slightly more but both should be ok</td></tr>
  <tr class="msg" id="L152"><td class="ts-cell"><a class="ts" href="#L152" alt="Mon Feb 29 2016 10:39:28 GMT-0800 (Pacific Standard Time)">18:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">Heading upstairs from lunch now, we can indeed probably land this today</td></tr>
  <tr class="msg" id="L153"><td class="ts-cell"><a class="ts" href="#L153" alt="Mon Feb 29 2016 10:39:31 GMT-0800 (Pacific Standard Time)">18:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: or &quot;I.e., they support being cloned across ...&quot;</td></tr>
  <tr class="msg" id="L154"><td class="ts-cell"><a class="ts" href="#L154" alt="Mon Feb 29 2016 10:40:04 GMT-0800 (Pacific Standard Time)">18:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">Yeah although I prefer to start sentences with &quot;that is&quot; instead of &quot;I.e.&quot;</td></tr>
  <tr class="msg" id="L155"><td class="ts-cell"><a class="ts" href="#L155" alt="Mon Feb 29 2016 10:40:47 GMT-0800 (Pacific Standard Time)">18:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">sure</td></tr>
  <tr class="msg" id="L156"><td class="ts-cell"><a class="ts" href="#L156" alt="Mon Feb 29 2016 10:47:46 GMT-0800 (Pacific Standard Time)">18:47</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-10" title=TabAtkins@irc>TabAtkins</span>&gt;</div></td><td class="msg-cell">Domenic: Well Bikeshed has the same issue, so I&#039;m still filing an issue on myself. ^_^</td></tr>
  <tr class="msg" id="L157"><td class="ts-cell"><a class="ts" href="#L157" alt="Mon Feb 29 2016 10:59:15 GMT-0800 (Pacific Standard Time)">18:59</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: thanks so much for all the reviewing</td></tr>
  <tr class="msg" id="L158"><td class="ts-cell"><a class="ts" href="#L158" alt="Mon Feb 29 2016 10:59:33 GMT-0800 (Pacific Standard Time)">18:59</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">annevk: thanks for doing this!! it&#039;s so much nicer.</td></tr>
  <tr class="msg" id="L159"><td class="ts-cell"><a class="ts" href="#L159" alt="Mon Feb 29 2016 10:59:51 GMT-0800 (Pacific Standard Time)">18:59</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=annevk@irc>annevk</span>&gt;</div></td><td class="msg-cell">Domenic: ECMAScript prose is rather addictive</td></tr>
  <tr class="msg" id="L160"><td class="ts-cell"><a class="ts" href="#L160" alt="Mon Feb 29 2016 11:00:12 GMT-0800 (Pacific Standard Time)">19:00</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">You can close <a href="https://github.com/whatwg/html/issues/575">https://github.com/whatwg/html/issues/575</a> and <a href="https://github.com/whatwg/html/issues/557">https://github.com/whatwg/html/issues/557</a> now if you want :)</td></tr>
  <tr class="msg" id="L161"><td class="ts-cell"><a class="ts" href="#L161" alt="Mon Feb 29 2016 11:00:21 GMT-0800 (Pacific Standard Time)">19:00</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">haha yeah I do like the precision</td></tr>
  <tr class="msg" id="L162"><td class="ts-cell"><a class="ts" href="#L162" alt="Mon Feb 29 2016 11:21:35 GMT-0800 (Pacific Standard Time)">19:21</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-11" title=wanderview_pto@irc>wanderview_pto</span>&gt;</div></td><td class="msg-cell">how many times should I write the same function before I propose a convenience version in the spec?</td></tr>
  <tr class="msg" id="L163"><td class="ts-cell"><a class="ts" href="#L163" alt="Mon Feb 29 2016 11:21:35 GMT-0800 (Pacific Standard Time)">19:21</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-11" title=wanderview_pto@irc>wanderview_pto</span>&gt;</div></td><td class="msg-cell">in this case, a promise-returning service worker wait_for_state()</td></tr>
  <tr class="msg" id="L164"><td class="ts-cell"><a class="ts" href="#L164" alt="Mon Feb 29 2016 11:24:07 GMT-0800 (Pacific Standard Time)">19:24</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">That sounds like a function very useful for tests and not that useful for other code?</td></tr>
  <tr class="msg" id="L165"><td class="ts-cell"><a class="ts" href="#L165" alt="Mon Feb 29 2016 11:27:13 GMT-0800 (Pacific Standard Time)">19:27</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-11" title=wanderview_pto@irc>wanderview_pto</span>&gt;</div></td><td class="msg-cell">ok</td></tr>
  <tr class="msg" id="L166"><td class="ts-cell"><a class="ts" href="#L166" alt="Mon Feb 29 2016 11:30:11 GMT-0800 (Pacific Standard Time)">19:30</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-1" title=Domenic@irc>Domenic</span>&gt;</div></td><td class="msg-cell">I dunno, are you writing tests or demos?</td></tr>
  <tr class="msg" id="L167"><td class="ts-cell"><a class="ts" href="#L167" alt="Mon Feb 29 2016 11:31:12 GMT-0800 (Pacific Standard Time)">19:31</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-11" title=wanderview_pto@irc>wanderview_pto</span>&gt;</div></td><td class="msg-cell">tests</td></tr>
  <tr class="msg" id="L168"><td class="ts-cell"><a class="ts" href="#L168" alt="Mon Feb 29 2016 11:33:22 GMT-0800 (Pacific Standard Time)">19:33</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=jsbell@irc>jsbell</span>&gt;</div></td><td class="msg-cell">tests + demos for me; I&#039;d love that too.</td></tr>
  <tr class="msg" id="L169"><td class="ts-cell"><a class="ts" href="#L169" alt="Mon Feb 29 2016 11:38:08 GMT-0800 (Pacific Standard Time)">19:38</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=jsbell@irc>jsbell</span>&gt;</div></td><td class="msg-cell">wanderview_pto: I think we talked about that very early on, maybe decided not to because the &quot;sw-as-progressive-enhancement&quot; scenario shouldn&#039;t use it.</td></tr>
  <tr class="msg" id="L170"><td class="ts-cell"><a class="ts" href="#L170" alt="Mon Feb 29 2016 11:39:31 GMT-0800 (Pacific Standard Time)">19:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=jsbell@irc>jsbell</span>&gt;</div></td><td class="msg-cell">But still, +1</td></tr>
  <tr class="msg" id="L171"><td class="ts-cell"><a class="ts" href="#L171" alt="Mon Feb 29 2016 11:39:53 GMT-0800 (Pacific Standard Time)">19:39</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-11" title=wanderview_pto@irc>wanderview_pto</span>&gt;</div></td><td class="msg-cell">jsbell: I dunno... I can think of progressive use cases... something like navigator.serviceWorker.register(script).then(swr =&gt; swr.installing.waitFor(&#039;active&#039;)).then(_ =&gt; enableOfflineUI())</td></tr>
  <tr class="msg" id="L172"><td class="ts-cell"><a class="ts" href="#L172" alt="Mon Feb 29 2016 11:40:06 GMT-0800 (Pacific Standard Time)">19:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-11" title=wanderview_pto@irc>wanderview_pto</span>&gt;</div></td><td class="msg-cell">you only get the offline UI if you can register a service worker that goes acitve</td></tr>
  <tr class="msg" id="L173"><td class="ts-cell"><a class="ts" href="#L173" alt="Mon Feb 29 2016 11:40:10 GMT-0800 (Pacific Standard Time)">19:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-12" title=jsbell@irc>jsbell</span>&gt;</div></td><td class="msg-cell">Right.</td></tr>
  <tr class="msg" id="L174"><td class="ts-cell"><a class="ts" href="#L174" alt="Mon Feb 29 2016 11:40:18 GMT-0800 (Pacific Standard Time)">19:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-11" title=wanderview_pto@irc>wanderview_pto</span>&gt;</div></td><td class="msg-cell">although I guess for &#039;active&#039; we already have the .ready promise</td></tr>
  <tr class="msg" id="L175"><td class="ts-cell"><a class="ts" href="#L175" alt="Mon Feb 29 2016 11:40:40 GMT-0800 (Pacific Standard Time)">19:40</a></td><td class="nick-cell"><div class="m-ov">&lt;<span class="nick nick-11" title=wanderview_pto@irc>wanderview_pto</span>&gt;</div></td><td class="msg-cell">I find the .ready promise a bit weird, tbh</td></tr>

</tbody></table></div></div></div></body>
